Add conversion

+++ impl/jamtemplate/common/tilemap/tilemap_collisions.cpp
@@ -29,25 +29,29 @@ void TilemapCollisions::refineColliders(float size)
     if (m_rects.empty()) {
         return;
     }
-    int const xmin = std::min_element(m_rects.cbegin(), m_rects.cend(),
-                         [](auto const& r1, auto const& r2) { return r1.left < r2.left; })
-                         ->left
-        / size;
+    int const xmin
+        = static_cast<int>(std::min_element(m_rects.cbegin(), m_rects.cend(),
+                               [](auto const& r1, auto const& r2) { return r1.left < r2.left; })
+                               ->left
+            / size);
 
-    int const ymin = std::min_element(m_rects.cbegin(), m_rects.cend(),
-                         [](auto const& r1, auto const& r2) { return r1.top < r2.top; })
-                         ->top
-        / size;
+    int const ymin
+        = static_cast<int>(std::min_element(m_rects.cbegin(), m_rects.cend(),
+                               [](auto const& r1, auto const& r2) { return r1.top < r2.top; })
+                               ->top
+            / size);
 
-    int const xmax = std::max_element(m_rects.cbegin(), m_rects.cend(),
-                         [](auto const& r1, auto const& r2) { return r1.left < r2.left; })
-                         ->left
-        / size;
+    int const xmax
+        = static_cast<int>(std::max_element(m_rects.cbegin(), m_rects.cend(),
+                               [](auto const& r1, auto const& r2) { return r1.left < r2.left; })
+                               ->left
+            / size);
 
-    int const ymax = std::max_element(m_rects.cbegin(), m_rects.cend(),
-                         [](auto const& r1, auto const& r2) { return r1.top < r2.top; })
-                         ->top
-        / size;
+    int const ymax
+        = static_cast<int>(std::max_element(m_rects.cbegin(), m_rects.cend(),
+                               [](auto const& r1, auto const& r2) { return r1.top < r2.top; })
+                               ->top
+            / size);
 
     std::vector<int> lookup;
     lookup.resize((xmax + 1) * (ymax + 1));
